将B树整合到查询中；将逻辑查询计划和物理查询计划分离

+++ src/main/java/db/file/BTree/BTreeHeaderPage.java
@@ -48,14 +48,14 @@ public class BTreeHeaderPage implements Page {
 		// Read the next and prev pointers
 		try {
 			Field f = Type.INT_TYPE.parse(dis);
-			this.nextPage = ((IntField) f).getValue();
+			this.nextPage = (int)((IntField) f).getValue();
 		} catch (java.text.ParseException e) {
 			e.printStackTrace();
 		}
 
 		try {
 			Field f = Type.INT_TYPE.parse(dis);
-			this.prevPage = ((IntField) f).getValue();
+			this.prevPage = (int)((IntField) f).getValue();
 		} catch (java.text.ParseException e) {
 			e.printStackTrace();
 		}
